--- SimG4Core/PhysicsLists/plugins/CMSModel.hh
@@ -1,14 +1,0 @@
-#ifndef SimG4Core_PhysicsLists_CMSModel_H
-#define SimG4Core_PhysicsLists_CMSModel_H
-
-#include "SimG4Core/Physics/interface/PhysicsList.h"
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-
-class CMSModel: public PhysicsList {
-
-public:
-  CMSModel(G4LogicalVolumeToDDLogicalPartMap& map, const HepPDT::ParticleDataTable * table_, sim::ChordFinderSetter *chordFinderSetter_, const edm::ParameterSet & p);
-};
-
-#endif
-
