+++ SimG4Core/PhysicsLists/plugins/FTFPCMS_BERT_EML.cc
@@ -1,4 +1,4 @@
-#include "FTFPCMS_BERT_EML.hh"
+#include "FTFPCMS_BERT_EML.h"
 #include "SimG4Core/PhysicsLists/interface/CMSEmStandardPhysics.h"
 #include "SimG4Core/PhysicsLists/interface/CMSMonopolePhysics.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
