--- SimG4Core/PhysicsLists/plugins/FTFPCMS_BERT_EML95.hh
@@ -1,16 +1,0 @@
-#ifndef SimG4Core_PhysicsLists_FTFPCMS_BERT_EML95_H
-#define SimG4Core_PhysicsLists_FTFPCMS_BERT_EML95_H
-
-#include "SimG4Core/Physics/interface/PhysicsList.h"
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-
-class FTFPCMS_BERT_EML95: public PhysicsList {
-
-public:
-  FTFPCMS_BERT_EML95(G4LogicalVolumeToDDLogicalPartMap& map, const HepPDT::ParticleDataTable * table_, sim::ChordFinderSetter *chordFinderSetter_, const edm::ParameterSet & p);
-};
-
-#endif
-
-
-
