+++ SimG4Core/PhysicsLists/plugins/FTFPCMS_BERT_HP_EML.cc
@@ -1,4 +1,4 @@
-#include "FTFPCMS_BERT_HP_EML.hh"
+#include "FTFPCMS_BERT_HP_EML.h"
 #include "SimG4Core/PhysicsLists/interface/CMSEmStandardPhysicsXS.h"
 #include "SimG4Core/PhysicsLists/interface/CMSMonopolePhysics.h"
 #include "SimG4Core/PhysicsLists/interface/CMSThermalNeutrons.h"
