+++ SimG4Core/PhysicsLists/plugins/QGSPCMS_BERT.cc
@@ -1,4 +1,4 @@
-#include "QGSPCMS_BERT.hh"
+#include "QGSPCMS_BERT.h"
 #include "SimG4Core/PhysicsLists/interface/CMSMonopolePhysics.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
