--- SimG4Core/PhysicsLists/plugins/QGSPCMS_BERT_EMLSYNC.cc
@@ -1,71 +1,0 @@
-#include "QGSPCMS_BERT_EMLSYNC.hh"
-#include "SimG4Core/PhysicsLists/interface/CMSEmStandardPhysics95msc93.h"
-#include "SimG4Core/PhysicsLists/interface/CMSMonopolePhysics.h"
-#include "FWCore/MessageLogger/interface/MessageLogger.h"
-
-#include "G4DecayPhysics.hh"
-#include "G4EmExtraPhysics.hh"
-#include "G4IonPhysics.hh"
-#include "G4StoppingPhysics.hh"
-#include "G4HadronElasticPhysics.hh"
-#include "G4NeutronTrackingCut.hh"
-#include "G4HadronicProcessStore.hh"
-
-#include "G4DataQuestionaire.hh"
-#include "G4HadronPhysicsQGSP_BERT.hh"
-
-QGSPCMS_BERT_EMLSYNC::QGSPCMS_BERT_EMLSYNC(G4LogicalVolumeToDDLogicalPartMap& map, 
-			   const HepPDT::ParticleDataTable * table_,
-			   sim::ChordFinderSetter *chordFinderSetter_, 
-			   const edm::ParameterSet & p) : PhysicsList(map, table_, chordFinderSetter_, p) {
-
-  G4DataQuestionaire it(photon);
-  
-  int  ver     = p.getUntrackedParameter<int>("Verbosity",0);
-  bool emPhys  = p.getUntrackedParameter<bool>("EMPhysics",true);
-  bool hadPhys = p.getUntrackedParameter<bool>("HadPhysics",true);
-  bool tracking= p.getParameter<bool>("TrackingCut");
-  edm::LogInfo("PhysicsList") << "You are using the simulation engine: "
-			      << "QGSP_BERT_EMLSYNC \n Flags for EM Physics "
-			      << emPhys << ", for Hadronic Physics "
-			      << hadPhys << " and tracking cut " << tracking;
-
-  if (emPhys) {
-    // EM Physics
-    RegisterPhysics( new CMSEmStandardPhysics95msc93("EM standard msc93",ver,""));
-
-    // Synchroton Radiation & GN Physics
-    G4EmExtraPhysics* extra = new G4EmExtraPhysics(ver);
-    G4String yes = "ON";
-    extra->Synch(yes);
-    RegisterPhysics(extra);
-  }
-
-  // Decays
-  this->RegisterPhysics( new G4DecayPhysics(ver) );
-
-  if (hadPhys) {
-    G4HadronicProcessStore::Instance()->SetVerbose(ver);
-
-    // Hadron Elastic scattering
-    RegisterPhysics( new G4HadronElasticPhysics(ver));
-
-    // Hadron Physics
-    RegisterPhysics(  new G4HadronPhysicsQGSP_BERT(ver));
-
-    // Stopping Physics
-    RegisterPhysics( new G4StoppingPhysics(ver));
-
-    // Ion Physics
-    RegisterPhysics( new G4IonPhysics(ver));
-
-    // Neutron tracking cut
-    if (tracking) {
-      RegisterPhysics( new G4NeutronTrackingCut(ver));
-    }
-  }
-
-  // Monopoles
-  RegisterPhysics( new CMSMonopolePhysics(table_,chordFinderSetter_,p));
-}
-
