--- SimG4Core/PhysicsLists/plugins/QGSPCMS_BERT_EMLSYNC.hh
@@ -1,15 +1,0 @@
-#ifndef SimG4Core_PhysicsLists_QGSPCMS_BERT_EMLSYNC_H
-#define SimG4Core_PhysicsLists_QGSPCMS_BERT_EMLSYNC_H
- 
-#include "SimG4Core/Physics/interface/PhysicsList.h"
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
- 
-class QGSPCMS_BERT_EMLSYNC: public PhysicsList {
-
-public:
-  QGSPCMS_BERT_EMLSYNC(G4LogicalVolumeToDDLogicalPartMap& map, const HepPDT::ParticleDataTable * table_, sim::ChordFinderSetter *chordFinderSetter_, const edm::ParameterSet & p);
-};
-
-#endif
-
-
