+++ SimG4Core/PhysicsLists/plugins/QGSPCMS_FTFP_BERT_EMY.cc
@@ -1,4 +1,4 @@
-#include "QGSPCMS_FTFP_BERT_EMY.hh"
+#include "QGSPCMS_FTFP_BERT_EMY.h"
 #include "SimG4Core/PhysicsLists/interface/CMSMonopolePhysics.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
