+++ SimG4Core/PhysicsLists/plugins/module.cc
@@ -1,26 +1,26 @@
 #include "SimG4Core/Physics/interface/PhysicsListFactory.h"
 
-#include "DummyPhysics.hh"
-#include "FTFCMS_BIC.hh"
-#include "FTFPCMS_BERT.hh"
-#include "FTFPCMS_BERT_ATL_EMM.hh"
-#include "FTFPCMS_BERT_EML.hh"
-#include "FTFPCMS_BERT_EMM.hh"
-#include "FTFPCMS_BERT_HP_EML.hh"
-#include "FTFPCMS_BERT_XS_EML.hh"
-#include "FTFPCMS_BERT_EMV.hh"
-#include "FTFPCMS_BERT_EMY.hh"
-#include "FTFPCMS_BERT_EMZ.hh"
-#include "QBBCCMS.hh"
-#include "QGSPCMS_BERT.hh"
-#include "QGSPCMS_BERT_EML.hh"
-#include "QGSPCMS_BERT_EMV.hh"
-#include "QGSPCMS_BERT_HP_EML.hh"
-#include "QGSPCMS_FTFP_BERT.hh"
-#include "QGSPCMS_FTFP_BERT_EML.hh"
-#include "QGSPCMS_FTFP_BERT_EMM.hh"
-#include "QGSPCMS_FTFP_BERT_EMY.hh"
-#include "QGSPCMS_FTFP_BERT_EMZ.hh"
+#include "DummyPhysics.h"
+#include "FTFCMS_BIC.h"
+#include "FTFPCMS_BERT.h"
+#include "FTFPCMS_BERT_ATL_EMM.h"
+#include "FTFPCMS_BERT_EML.h"
+#include "FTFPCMS_BERT_EMM.h"
+#include "FTFPCMS_BERT_HP_EML.h"
+#include "FTFPCMS_BERT_XS_EML.h"
+#include "FTFPCMS_BERT_EMV.h"
+#include "FTFPCMS_BERT_EMY.h"
+#include "FTFPCMS_BERT_EMZ.h"
+#include "QBBCCMS.h"
+#include "QGSPCMS_BERT.h"
+#include "QGSPCMS_BERT_EML.h"
+#include "QGSPCMS_BERT_EMV.h"
+#include "QGSPCMS_BERT_HP_EML.h"
+#include "QGSPCMS_FTFP_BERT.h"
+#include "QGSPCMS_FTFP_BERT_EML.h"
+#include "QGSPCMS_FTFP_BERT_EMM.h"
+#include "QGSPCMS_FTFP_BERT_EMY.h"
+#include "QGSPCMS_FTFP_BERT_EMZ.h"
 
 DEFINE_PHYSICSLIST(DummyPhysics);
 typedef FTFCMS_BIC FTF_BIC;
